requirements.txt
monai==0.5.0
torch
gdown
tqdm
itk
nibabel
notebook
python-dateutil
rootpath
scikit-image
scipy
SimpleITK
six
tensorboard
mlflow
pytorch-ignite==0.4.4
psutil
nnunet
numpy
argparse
pandas
matplotlib
PyYAML
pydantic
sklearn
scikit-learn
